Improved code re-use
