+++ include/aruco_detector/detector.h
@@ -13,6 +13,7 @@
 #include <geometry_msgs/Point.h>
 #include <geometry_msgs/PoseStamped.h>
 #include <sensor_msgs/CameraInfo.h>
+#include <maara_msgs/StereoCameraInfo.h>
 
 using namespace cv;
 using namespace std;
@@ -21,7 +22,12 @@ class MarkerDetector {
 public:
     MarkerDetector();
 
-    std::map<int, geometry_msgs::Pose> processImages(Mat image,  sensor_msgs::CameraInfo camera_info, double marker_size,  bool display);
+    //Stereo Detection
+    std::map<int, geometry_msgs::Pose> processImages(Mat left_image, Mat right_image, maara_msgs::StereoCameraInfo stereo_info, double marker_size, bool display);
+    //Single Image
+    std::map<int, geometry_msgs::Pose> processImages(Mat image, sensor_msgs::CameraInfo camera_info, double marker_size, bool display);
+    //Depth Detection
+    std::map<int, geometry_msgs::Pose> processImage(Mat image, Mat depth_image, sensor_msgs::CameraInfo camera_info, bool display);
 
 private:
